moved to customDataModelUtils for all composites
