+++ packages/react-composites/src/composites/CallComposite/components/MediaGallery.tsx
@@ -18,7 +18,7 @@ import { useSelector } from '../hooks/useSelector';
 import { localVideoCameraCycleButtonSelector } from '../selectors/LocalVideoTileSelector';
 /* @conditional-compile-remove(call-with-chat-composite) @conditional-compile-remove(local-camera-switcher) */
 import { LocalVideoCameraCycleButton } from '@internal/react-components';
-import { useCustomAvatarPersonaData } from '../utils';
+import { useCustomAvatarPersonaData } from '../../common/CustomDataModelUtils';
 
 const VideoGalleryStyles = {
   root: {
