+++ packages/react-composites/src/composites/CallComposite/utils/Utils.ts
@@ -5,8 +5,6 @@ import { CallAdapterState, CallCompositePage } from '../adapter/CallAdapter';
 import { _isInCall, _isPreviewOn, _isInLobbyOrConnecting } from '@internal/calling-component-bindings';
 import { CallControlOptions } from '../types/CallControlOptions';
 import { CallState } from '@internal/calling-stateful-client';
-import { AvatarPersonaData, AvatarPersonaDataCallback } from '../../common/AvatarPersona';
-import React, { useEffect } from 'react';
 
 const ACCESS_DENIED_TEAMS_MEETING_SUB_CODE = 5854;
 const REMOVED_FROM_CALL_SUB_CODES = [5000, 5300];
@@ -185,63 +183,4 @@ export const computeVariant = (
   } else {
     return 'NO_STATE';
   }
-};
-
-/**
- * Hook to override the avatar persona data of the users
- * - Calls {@link AvatarPersonaDataCallback} on each render for provided `userIds`
- * - Returns an array of the same length as `userIds`. Entries in the array may be undefined if there is no data to return.
- * @private
- */
-export const useCustomAvatarPersonaData = (
-  userIds: (string | undefined)[],
-  callback?: AvatarPersonaDataCallback
-): (AvatarPersonaData | undefined)[] => {
-  const [data, setData] = React.useState<(AvatarPersonaData | undefined)[]>(Array(userIds.length).fill(undefined));
-  useEffect(() => {
-    (async () => {
-      const newData = await Promise.all(
-        userIds.map(async (userId: string | undefined) => {
-          if (!userId || !callback) {
-            return undefined;
-          }
-          return await callback(userId);
-        })
-      );
-      if (shouldUpdate(data, newData)) {
-        setData(newData);
-      }
-    })();
-  });
-  return data;
-};
-
-/**
- * Function to determine if there is new user avatar data in current render pass.
- * @param currentData current set set of avatar persona data present from the custom settings from the previous render
- * @param newData new set of avatar persona data after a run of {@link useCustomAvatarPersonData}
- * @returns Boolean whether there is new avatar persona data present.
- * @private
- */
-const shouldUpdate = (
-  currentData: (AvatarPersonaData | undefined)[],
-  newData: (AvatarPersonaData | undefined)[]
-): boolean => {
-  if (currentData.length !== newData.length) {
-    return true;
-  }
-  return currentData.some((data, i) => avatarDeepEqual(data, newData[i]));
-};
-
-/**
- * @private
- */
-const avatarDeepEqual = (currentData?: AvatarPersonaData, newData?: AvatarPersonaData): boolean => {
-  return currentData?.text !== newData?.text &&
-    currentData?.imageUrl !== newData?.imageUrl &&
-    currentData?.initialsColor !== newData?.initialsColor &&
-    currentData?.imageInitials !== newData?.imageInitials &&
-    currentData?.initialsTextColor !== newData?.initialsTextColor
-    ? true
-    : false;
 };
